Refactors vertex shader compiler management
