+++ examples/shader-generator/CustomCompiler.js
@@ -11,9 +11,6 @@
         osgShader.Compiler.apply( this, arguments );
     };
 
-    CustomCompiler.validAttributeType = osgShader.Compiler.validAttributeType.slice();
-    CustomCompiler.validAttributeType.push( 'Negatif', 'Ramp' );
-
     CustomCompiler.prototype = osg.objectInherit( osgShader.Compiler.prototype, {
