+++ examples/shader-generator/main.js
@@ -102,6 +102,10 @@
             var shaderGenerator = new osgShader.ShaderGenerator();
             shaderGenerator.setShaderCompiler( CustomCompiler );
 
+            var accepted = shaderGenerator.getAcceptAttributeTypes();
+            accepted.add( 'Negatif' );
+            accepted.add( 'Ramp' );
+
             // get or create instance of ShaderGeneratorProxy
             var shaderGeneratorProxy = this._viewer.getState().getShaderGeneratorProxy();
             shaderGeneratorProxy.addShaderGenerator( 'custom', shaderGenerator );
