+++ examples/shader-generator/negatif.js
@@ -10,7 +10,8 @@
         osg.StateAttribute.call( this );
         this._attributeEnable = false;
     };
-    osg.createPrototypeStateAttribute( NegatifAttribute, osg.objectInherit( osg.StateAttribute.prototype, {
+
+    NegatifAttribute.prototype = osg.objectLibraryClass( osg.objectInherit( osg.StateAttribute.prototype, {
         attributeType: 'Negatif',
 
         cloneType: function () {
