+++ sources/osgShadow/ShadowCastCompiler.js
@@ -5,17 +5,17 @@ var Compiler = require( 'osgShader/Compiler' );
 
 var CompilerShadowCast = function () {
     Compiler.apply( this, arguments );
-    this._isVertexColored = false;
-    this._isLighted = false;
 };
 
 CompilerShadowCast.prototype = MACROUTILS.objectInherit( Compiler.prototype, {
     getCompilerName: function () {
         return 'ShadowCast';
     },
+
     getFragmentShaderName: function () {
         return this.getCompilerName();
     },
+
     initAttributes: function () {
         var attributes = this._attributes;
 
@@ -34,19 +34,11 @@ CompilerShadowCast.prototype = MACROUTILS.objectInherit( Compiler.prototype, {
             }
         }
     },
+
     registerTextureAttributes: function () {},
-    registerTextureShadow: function () {},
 
-    // Fast Path, only Depth
-    declareVertexTransforms: function ( glPosition ) {
-        this.declareTransformWithEyeSpace( glPosition );
-    },
-
-    // Depth Shadow Map Casted from Light POV
-    // Depth encoded in color buffer
-    createShadowCastDepth: function () {
-
-        var frag = this.createVariable( 'vec4' );
+    // Depth Shadow Map Casted from Light POV Depth encoded in color buffer
+    createShadowCastDepth: function ( out ) {
 
         this.getNode( 'ShadowCast' ).setShadowCastAttribute( this._shadowCastAttribute ).inputs( {
 
@@ -57,22 +49,18 @@ CompilerShadowCast.prototype = MACROUTILS.objectInherit( Compiler.prototype, {
 
         } ).outputs( {
 
-            color: frag
+            color: out
 
         } );
 
-        return frag;
+        return out;
     },
 
     // encapsulate for easier overwrite by user defined compiler
-    // that would inherint from this compiler
-    // Do not merge with above method
+    // that would inherint from this compiler Do not merge with above method
     createFragmentShaderGraph: function () {
-
-        var depthFrag = this.createShadowCastDepth();
         var frag = this.getNode( 'glFragColor' );
-        this.getNode( 'SetFromNode' ).inputs( depthFrag ).outputs( frag );
-        return [ frag ];
+        return [ this.createShadowCastDepth( frag ) ];
     }
 
 } );
