+++ sources/osgUtil/DisplayNormalVisitor.js
@@ -21,7 +21,6 @@ var UpdateMorph = require( 'osgAnimation/UpdateMorph' );
 ////////////////////////
 var CompilerOffsetNormal = function () {
     Compiler.apply( this, arguments );
-    this._isVertexColored = false;
 };
 
 CompilerOffsetNormal.prototype = MACROUTILS.objectInherit( Compiler.prototype, {
@@ -42,20 +41,19 @@ CompilerOffsetNormal.prototype = MACROUTILS.objectInherit( Compiler.prototype, {
         return [ frag ];
     },
     _getDirectionVec: function () {
-        return this.getOrCreateNormalAttribute();
+        return this.getOrCreateWorldNormal();
     },
-    getOrCreateVertexAttribute: function () {
+    getOrCreateWorldPosition: function () {
         var vertexOffset = this.getVariable( 'vertexOffset' );
         if ( vertexOffset ) return vertexOffset;
 
         vertexOffset = this.createVariable( 'vec3', 'vertexOffset' );
 
-        // normalize len (divide scale of worldmat, don't work for non uniform scale, but work enough for debug)
-        var str = '%out = %offset == 1.0 ? %vertex + normalize(%direction.xyz) * %scale / length(%world[0].xyz) : %vertex;';
+        var str = '%out = %offset == 1.0 ? %vertex + normalize(%direction.xyz) * %scale: %vertex;';
         this.getNode( 'InlineCode' ).code( str ).inputs( {
             offset: this.getOrCreateAttribute( 'float', 'Offset' ),
             direction: this._getDirectionVec(),
-            vertex: Compiler.prototype.getOrCreateVertexAttribute.call( this ),
+            vertex: Compiler.prototype.getOrCreateWorldPosition.call( this ),
             scale: this.getOrCreateUniform( 'float', 'uScale' ),
             world: this.getOrCreateUniform( 'mat4', 'ModelWorldMatrix' )
         } ).outputs( {
@@ -64,7 +62,20 @@ CompilerOffsetNormal.prototype = MACROUTILS.objectInherit( Compiler.prototype, {
 
         return vertexOffset;
     },
-    declareVertexTransforms: Compiler.prototype.declareVertexTransformShadeless
+    getOrCreateViewVertex: function () {
+        var out = this._variables.FragEyeVector;
+        if ( out && !out.isEmpty() ) return out;
+        out = this._varyings.FragEyeVector || this.createVariable( 'vec4', 'FragEyeVector' );
+
+        this.getNode( 'MatrixMultPosition' ).inputs( {
+            matrix: this.getOrCreateUniform( 'mat4', 'ViewMatrix' ),
+            vec: this.getOrCreateWorldPosition()
+        } ).outputs( {
+            vec: out
+        } );
+
+        return out;
+    }
 } );
 
 var ShaderGeneratorCompilerOffsetNormal = function () {
@@ -85,7 +96,7 @@ CompilerOffsetTangent.prototype = MACROUTILS.objectInherit( CompilerOffsetNormal
         return 'CompilerOffsetTangent';
     },
     _getDirectionVec: function () {
-        return this.getOrCreateTangentAttribute();
+        return this.getOrCreateWorldTangent();
     }
 } );
